restructure error and remove unused code
