+++ near/contract/src/error.rs
